+++ lib/contentful/management/resource_builder.rb
@@ -6,6 +6,7 @@ require_relative 'array'
 require_relative 'link'
 require_relative 'space'
 require_relative 'space_membership'
+require_relative 'organization'
 require_relative 'content_type'
 require_relative 'asset'
 require_relative 'webhook'
@@ -29,6 +30,7 @@ module Contentful
       DEFAULT_RESOURCE_MAPPING = {
         'Space' => Contentful::Management::Space,
         'SpaceMembership' => Contentful::Management::SpaceMembership,
+        'Organization' => Contentful::Management::Organization,
         'ContentType' => Contentful::Management::ContentType,
         'Entry' => :find_entry_class,
         'Asset' => Contentful::Management::Asset,